requirements.txt
os
json
shutil
tkinter
python-docx
